try fix: run lora path in fp16 and cast back to fp16

+++ loraprune/lora.py
@@ -307,10 +307,12 @@ class Linear(nn.Linear, LoraLayer):
             result = F.linear(x, transpose(self.weight, self.fan_in_fan_out), bias=self.bias)
         elif self.r > 0 and not self.merged:
             result = F.linear(x, transpose(self.weight, self.fan_in_fan_out), bias=self.bias)
+            # Compute lora path in fp32: if upstream activations overflow fp16 (inf),
+            # casting lora_B=0 to fp16 and doing F.linear(inf, 0) = NaN corrupts training.
             lora_output = F.linear(
-                F.linear(self.lora_dropout(x), self.lora_A.weight.to(x.dtype)),
-                self.lora_B.weight.to(x.dtype)
-            ) * self.scaling
+                F.linear(self.lora_dropout(x).float(), self.lora_A.weight),
+                self.lora_B.weight
+            ).to(x.dtype) * self.scaling
             result = result + lora_output
         else:
             result = F.linear(x, transpose(self.weight, self.fan_in_fan_out), bias=self.bias)
